+++ src/main/java/insane96mcp/mobspropertiesrandomness/json/MPREquipment.java
@@ -104,7 +104,7 @@ public class MPREquipment implements IMPRObject, IMPRAppliable {
 		entity.setItemStackToSlot(equipmentSlotType, itemStack);
 
 		//Drop Chance
-		entity.setDropChance(equipmentSlotType, choosenItem.dropChance);
+		entity.setDropChance(equipmentSlotType, choosenItem.dropChance.getValue(entity, world));
 
 	}