+++ src/main/java/insane96mcp/mobspropertiesrandomness/setup/Strings.java
@@ -6,5 +6,6 @@ public class Strings {
 	public static class Tags {
 		public static final String PROCESSED = MobsPropertiesRandomness.RESOURCE_PREFIX + "processed";
 		public static final String SPAWNED_FROM_SPAWNER = MobsPropertiesRandomness.RESOURCE_PREFIX + "spawned_from_spawner";
+		public static final String SPAWNED_FROM_STRUCTURE = MobsPropertiesRandomness.RESOURCE_PREFIX + "spawned_from_structure";
 	}
 }