+++ changelog.md
@@ -4,9 +4,10 @@
 * Reworked random enchantments  
   "random" is a new Object in the Enchantment Object. You no longer have to write "random" in the "id" key.  
   The RandomEnchantment object contains "allow_curses", "allow_treasures" and a list of possible enchantments. If the list is omitted, the old behaviour will apply (random enchantment out of all the possible ones for the item)
-* Added chance to attributes (much like potion effects). Also, health and follow range are now fixed for items too
-* Added a deepness property to WorldWhitelist object. This Range Object dictates the min and max Y position at which the property will be applied
-* Range and Modifiable value objects can now be set with just a number, instead of having to create an object with the "value" (or "min") key inside
+* Added `chance` to attributes (much like potion effects). Also, health and follow range are now fixed for items too
+* Added a `deepness` property to WorldWhitelist object. This Range Object dictates the min and max Y position at which the property will be applied
+* Added a `structure_behaviour` property to the mob. Much like `spawner_behaviour` will filter mobs spawned from structures. Valid values are NONE, STRUCTURE_ONLY, NATURAL_ONLY.
+* Range and Modifiable value objects can now be set with just a number, instead of having to create an object with the "value" (or "min") key inside.  
 * Items Drop Chance is now a Modifiable Value
 
 ## 3.0.5